refactor: reorganizes the plugin directory
This structure will make more sense in the light of the
other mediator components being introduced for the packages
repositories
